Adding charUtils
