+++ kernel/inc/charUtils.h
@@ -1,0 +1,12 @@
+#ifndef CHAR_UTILS_H
+#define CHAR_UTILS_H
+
+#include "intTypeDefs.h"
+
+bool isspace(char);
+
+bool isnum(char);
+
+int ctoi(char);
+
+#endif
