+++ kernel/inc/stringUtils.c
@@ -59,14 +59,6 @@ string strcat(string dest, string src)
     return dest;
 }
 
-bool isspace(char ch) {
-    return (ch == ' ') || (ch == '\t') || (ch == '\r') || (ch == '\n') || (ch == '\f') || (ch == '\0');
-}
-
-bool isnum(char ch) {
-    return (ch >= '0') && (ch <= '9');
-}
-
 string strTrim(string str)
 {
     uint16 len = 0;
@@ -134,23 +126,6 @@ string itos(int i)
         *--p = '-';
     }
     return p;
-}
-
-int ctoi(char c)
-{
-    switch(c) {
-    case '0': return 0;
-    case '1': return 1;
-    case '2': return 2;
-    case '3': return 3;
-    case '4': return 4;
-    case '5': return 5;
-    case '6': return 6;
-    case '7': return 7;
-    case '8': return 8;
-    case '9': return 9;
-    default: return -1;
-    }
 }
 
 int stoi(string s)
