+++ kernel/inc/stringUtils.h
@@ -1,6 +1,7 @@
 #ifndef STRINGUTILS_H
 #define STRINGUTILS_H
 
+#include "charUtils.h"
 #include "intTypeDefs.h"
 
 uint16 strlength(string);
@@ -15,13 +16,7 @@ string strcat(string, string);
 
 string strTrim(string);
 
-bool isspace(char);
-
-bool isnum(char);
-
 string itos(int);
-
-int ctoi(char);
 
 int stoi(string);
